requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
resend
